Whole bunch of changes.
Finally getting back in the swing of things after having my old OS canned...
But I should be on the road to getting this things groundwork done here soon.
Once that's finished I have two things that I want to do then I'll be able
to move on to other projects. FINALLY...